% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylostructure.R
\name{phylostructure}
\alias{phylostructure}
\title{Generating Node Relations}
\usage{
phylostructure(phylotree)
}
\arguments{
\item{phylotree}{A \code{phylo-class} object as in the \code{phyloseq} package. 
The tree has to be binary.}
}
\value{
A list containing the following components:
\item{\code{phylotree}}{Input object}
\item{\code{phylochildren}}{Numerical matrix of child node labels. The kth row 
contains the children node labels of node k. Leaf nodes have their their row 
vectors set to zero.}
\item{\code{phyloparent}}{Numerical vector of parent node labels. The kth element 
contains the parent node label of node k. Root node has its parent set to zero.}
\item{\code{descendant}}{Logical matrix of decendant leaves. The kth row 
shows whether the leaves are descendants of node k.}
}
\description{
\code{phylostructure} returns the neighboring nodes and descendant leaves 
for each internal node of a binary phylogenetic tree.
}
\details{
For each internal node (i.e. non-leaf node) in \code{phylotree}, 
this function calculates its parent node, child nodes and all of its descendant
leaves. The labeling of nodes are the same as in \code{phylotree$edge}, which is 
assumed to obey the following convention. Leaves are labeled from \code{1} to 
\code{ntaxa(phylotree)} and internal nodes are labeled from 
\code{ntaxa(phylotree) + 1} to \code{ntaxa(phylotree) + phylotree$Nnode}. The 
root label is \code{ntaxa(phylotree) + 1}. Parent node always has a lower label 
than its child internal nodes.
}
\examples{
library(ape)
set.seed(10)

## Analyze a random binary phylogenetic tree with 6 leaf nodes
pstrct <- phylostructure(rtree(6))
## Show children of node 7 (root):
pstrct$phylochildren[7, ]
## Show parent of node 1:
pstrct$phyloparent[1]
## List all leaf nodes under node 8:
which(pstrct$descendant[8, ])

}
\author{
Yunfan Tang
}

